+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This log is intended to keep track of package changes, including
 but not limited to API changes and file location changes. Minor behavioral
 changes may not be included if they are not expected to break existing code.
 
+## 0.5.0 (2023-10-31)
+
+- Update `onError` payloads to return a `OneSchemaError`
+- Use `onError` for launch errors
+- Emit HTTP errors from `onError`
+
 ## 0.4.8 (2023-10-25)
 
 - Stop `_initWithRetry` if the postMessage has been recieved by OneSchema app.